Move MenuList component to @gw2treasures/ui
