+++ apps/web/components/Format/FormatConfigDialog.tsx
@@ -3,7 +3,7 @@ import { Dialog } from '../Dialog/Dialog';
 import { Label } from '@gw2treasures/ui/components/Form/Label';
 import { Select } from '@gw2treasures/ui/components/Form/Select';
 import { useLanguage } from '../I18n/Context';
-import { MenuList } from '../MenuList/MenuList';
+import { MenuList } from '@gw2treasures/ui/components/Layout/MenuList';
 import { useFormatContext } from './FormatContext';
 import { FormatDate } from './FormatDate';
 import { FormatNumber } from './FormatNumber';
