+++ apps/web/components/ItemTable/ItemTable.client.tsx
@@ -9,7 +9,7 @@ import { Table } from '@gw2treasures/ui/components/Table/Table';
 import { DropDown } from '../DropDown/DropDown';
 import { Button, LinkButton } from '@gw2treasures/ui/components/Form/Button';
 import { Icon } from '@gw2treasures/ui';
-import { MenuList } from '../MenuList/MenuList';
+import { MenuList } from '@gw2treasures/ui/components/Layout/MenuList';
 import { encode } from 'gw2e-chat-codes';
 import { CopyButton } from '@gw2treasures/ui/components/Form/Buttons/CopyButton';
 import { Pagination } from '../Pagination/Pagination';
