+++ apps/web/components/ItemTable/ItemTableColumnsButton.tsx
@@ -4,7 +4,7 @@ import type { FC } from 'react';
 import { DropDown } from '../DropDown/DropDown';
 import { Button } from '@gw2treasures/ui/components/Form/Button';
 import { Checkbox } from '@gw2treasures/ui/components/Form/Checkbox';
-import { MenuList } from '../MenuList/MenuList';
+import { MenuList } from '@gw2treasures/ui/components/Layout/MenuList';
 import { Separator } from '@gw2treasures/ui/components/Layout/Separator';
 import { useItemTableContext } from './context';
 import { Icon } from '@gw2treasures/ui';
