+++ apps/web/components/Recipe/RecipeTable.tsx
@@ -12,7 +12,7 @@ import { Headline } from '@gw2treasures/ui/components/Headline/Headline';
 import { useLanguage } from '../I18n/Context';
 import { ItemLink, type ItemLinkProps } from '../Item/ItemLink';
 import { Separator } from '@gw2treasures/ui/components/Layout/Separator';
-import { MenuList } from '../MenuList/MenuList';
+import { MenuList } from '@gw2treasures/ui/components/Layout/MenuList';
 import { ShowMore } from '../ShowMore/ShowMore';
 import { Table } from '@gw2treasures/ui/components/Table/Table';
 import { type Discipline, DisciplineIcon } from './DisciplineIcon';
